frontend/requirements.txt
Flask==3.0.3
requests==2.31.0
python-dotenv==1.0.1
gunicorn==21.2.0
Werkzeug==3.0.3